+++ app/src/main/java/host/stjin/anonaddy/notifications/ActionReceiver.kt
@@ -18,6 +18,7 @@ class ActionReceiver : BroadcastReceiver() {
         const val STOP_UPDATE_CHECK = "stop_update_check"
         const val STOP_FAILED_DELIVERY_CHECK = "stop_failed_delivery_check"
         const val STOP_API_EXPIRY_CHECK = "stop_api_expiry_check"
+        const val STOP_SUBSCRIPTION_EXPIRY_CHECK = "stop_subscription_expiry_check"
         const val STOP_PERIODIC_BACKUPS = "stop_periodic_backups"
         const val DISABLE_WEAROS_QUICK_SETUP = "disable_wearos_quick_setup"
     }
@@ -62,6 +63,11 @@ class ActionReceiver : BroadcastReceiver() {
                 SettingsManager(false, context).putSettingsBool(SettingsManager.PREFS.NOTIFY_API_TOKEN_EXPIRY, false)
                 // Dismiss notification
                 notificationManager.cancel(NotificationHelper.API_KEY_EXPIRE_NOTIFICATION_ID)
+            }
+            NOTIFICATIONACTIONS.STOP_SUBSCRIPTION_EXPIRY_CHECK -> {
+                SettingsManager(false, context).putSettingsBool(SettingsManager.PREFS.NOTIFY_SUBSCRIPTION_EXPIRY, false)
+                // Dismiss notification
+                notificationManager.cancel(NotificationHelper.SUBSCRIPTION_EXPIRE_NOTIFICATION_ID)
             }
             NOTIFICATIONACTIONS.STOP_PERIODIC_BACKUPS -> {
                 SettingsManager(false, context).putSettingsBool(SettingsManager.PREFS.PERIODIC_BACKUPS, false)