+++ CHANGELOG.md
@@ -1,5 +1,25 @@
 # Changelog
 
+## [v4.6.0] - 2023-03-12
+
+### AnonAddy for Android
+
+##### Added
+
+- 💳 Get notified when your subscription is about to expire thanks to the new "Subscription expiry notification"-feature which can be enabled under **App settings->Features and integrations**
+
+##### Fixed/Improved
+
+- ✨ Fixed profile sheet not showing an alert when an update is available
+- ✨ Limit the amount of expiry notifications to 1 every day
+- ✨ Performance improvements and other things I might have forgot to make the app even smoother
+
+### AnonAddy for Android Wearables _(v1.1.4)_
+
+##### Fixed/Improved
+
+- ✨ Performance improvements and other things to make the app even smoother
+
 ## [v4.5.3] - 2023-02-25
 
 ### AnonAddy for Android
@@ -344,6 +364,7 @@
 ## [v2.0.3] - 2021-04-19
 
 ### Fixed/Improved
+
 - [IMPROVED] Updated libraries
 - [IMPROVED] Added "Use Reply-To Header For Replying" to AnonAddy settings
 - [IMPROVED] Overall app improvements
@@ -351,22 +372,27 @@
 ## [v2.0.2] - 2021-02-08
 
 ### Fixed/Improved
+
 - [IMPROVED] Updated shared domains
 
 ## [v2.0.1] - 2021-01-18
 
 ### Added
+
 - 📋 [NEW] Sending an email from an alias will now also copy the recipients to the clipboard
 
 ## [v2.0.0] - 2021-01-11
 
 ### Added
+
 - 🌐 [NEW] Get full control over how often background data such as widgets are refreshed
 - 👁️ [NEW] [APP EXCLUSIVE] Watch your aliases to get notified when it forwards a new email
 - ✉  [NEW] Quickly send emails from an alias
 - 🔤  [NEW] Added random characters alias format
 - 💳 [NEW] View the subscription end date
+
 ### Fixed/Improved
+
 - 🖌️ [IMPROVED] There are some major and minor UI improvements making the app look even better on phones, tablets and foldables
 - 🔐 [IMPROVED] The app just got even more secure as it now also requires authentication for actions being executed from the widget or intents
 - 🧈 [IMPROVED] Lot's of love and butter has gone into this release to make the app smoother and more lightweight than ever before. Especially the widget got a big improvement
@@ -375,12 +401,14 @@
 ## [v1.1.3] - 2020-12-08
 
 ### Fixed/Improved
+
 - 🐛 [BUG FIX] Deleted alias section is not collapsed by default
 - 🙋 [FEATURE REQUEST] Switch between panels by swiping left/right
 
 ## [v1.1.2] - 2020-11-21
 
 ### Fixed/Improved
+
 - 🖥️ Optimized UI for big screens
 - 🐛 [BUG FIX] Improved responsiveness when switching fragments
 - 🙋 [FEATURE REQUEST] Added link to gitlab page
@@ -389,7 +417,8 @@
 ## [v1.1.1] - 2020-10-24
 
 ### Fixed/Improved
--  Added the forgotten piece of the custom alias format >.<
+
+- Added the forgotten piece of the custom alias format >.<
 - Fixed crash at double Biometrics authentication prompt when dark mode was enabled
 
 ## [v1.1.0] - 2020-10-24
@@ -405,7 +434,7 @@
 - 💌 Show changelog on update
 - 🛠️ Added version check (for self hosted instances)
 
-
 ### Fixed/Improved
+
 - 🔎 Improved the search function
 - 🌟 UI improvements